Moves test to proper Spec at javadsl-api.
