+++ service/javadsl/api/src/test/java/com/lightbend/lagom/internal/api/NotPublicInterfaceService.java
@@ -1,7 +1,7 @@
 /*
  * Copyright (C) 2016 Lightbend Inc. <http://www.lightbend.com>
  */
-package com.lightbend.lagom.it.failures;
+package com.lightbend.lagom.internal.api;
 
 import com.lightbend.lagom.javadsl.api.Descriptor;
 import com.lightbend.lagom.javadsl.api.Service;
